Revert "[auto parallel] add softmax backward spmd rule (#59039)"
This reverts commit d86f6866464f4ca2c73dd8014bb4dad392cfff25.
