+++ paddle/phi/api/yaml/legacy_backward.yaml
@@ -661,7 +661,6 @@
   infer_meta :
     func : UnchangedInferMeta
     param : [out]
-    spmd_rule : SoftmaxGradInferSpmd
   kernel :
     func : softmax_grad
   composite : softmax_grad(out, out_grad, axis, x_grad)
