+++ paddle/phi/api/yaml/legacy_ops.yaml
@@ -1004,7 +1004,6 @@
   output : Tensor(out)
   infer_meta :
     func : SoftmaxInferMeta
-    spmd_rule : SoftmaxInferSpmd
   kernel :
     func : softmax
   inplace : (x -> out)
