+++ paddle/phi/infermeta/spmd_rules/softmax.h
@@ -31,8 +31,5 @@ SpmdInfo SoftmaxInferSpmdReverse(const DistMetaTensor& x,
                                  const DistMetaTensor& out,
                                  int axis);
 
-SpmdInfo SoftmaxGradInferSpmd(const DistMetaTensor& out,
-                              const DistMetaTensor& out_grad,
-                              int axis);
 }  // namespace distributed
 }  // namespace phi
